+++ backend/src/app.module.ts
@@ -21,20 +21,18 @@ import { APP_GUARD } from "@nestjs/core";
 import { RolesGuard } from "./auth/roles.guard";
 import { LikeModel } from "./post/models/like.model";
 import { DislikeModel } from "./post/models/dislike.model";
-import { MessageGateway } from "./chat/message.gateway";
-import { AlertGateway } from "./alert/alert.gateway";
-import { MessageController } from "./chat/message.controller";
-import { AlertController } from "./alert/alert.controller";
+import { MessageGateway } from "./message/message.gateway";
+import { MessageController } from "./message/message.controller";
+import { MessageModule } from "./message/message.module";
 
 @Module({
-  controllers: [MessageController, AlertController],
+  controllers: [MessageController],
   providers: [
     {
       provide: APP_GUARD,
       useClass: RolesGuard,
     },
     MessageGateway,
-    AlertGateway,
   ],
   imports: [
     ConfigModule.forRoot({
@@ -71,6 +69,7 @@ import { AlertController } from "./alert/alert.controller";
     PostModule,
     ProfileModule,
     FileModule,
+    MessageModule,
   ],
 })
 export class AppModule {}