+++ backend/src/message/models/message.model.ts
@@ -1,4 +1,4 @@
-import { BelongsTo, Column, DataType, HasMany, Model, Table } from "sequelize-typescript";
+import { BelongsTo, Column, DataType, Model, Table } from "sequelize-typescript";
 import { ApiProperty } from "@nestjs/swagger";
 import { UserModel } from "../../user/models/user.model";
 import { ConversationModel } from "./conversation.model";
@@ -17,9 +17,9 @@ export class MessageModel extends Model<MessageModel, MessageCreationAttr> {
   @Column({ type: DataType.STRING })
   message: string;
 
-  @HasMany(() => UserModel)
+  @BelongsTo(() => UserModel)
   user: UserModel[];
 
   @BelongsTo(() => ConversationModel)
-  conversations: ConversationModel[];
+  conversations: ConversationModel;
 }