+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Features:
 
 Backend:
+1. Added conversation service, model.
+2. Added user-conversation, active-conversation and message-conversation models
+
+# v2.5.1 (16.03.2022)
+
+## Features:
+
+Backend:
 1. Deleted alerts
 2. Added message module, model
 3. Changed User model